fix file renaming mix-up
