+++ lib/server.js
@@ -121,7 +121,8 @@ var SPDYProxy = function(options) {
   }
 
   function handleRequest(req, res) {
-    var socket = (req.method == 'CONNECT') ? res : res.socket;
+    var https = req.method == 'CONNECT';
+    var socket = https ? res : res.socket;
     console.log("%s:%s".yellow + " - %s - " + "stream ID: " + "%s".yellow + " - priority: " + "%s".yellow,
       socket.connection ? socket.connection.socket.remoteAddress : socket.socket.remoteAddress,
       socket.connection ? socket.connection.socket.remotePort : socket.socket.remotePort,
@@ -134,44 +135,51 @@ var SPDYProxy = function(options) {
     // want to pass this injected header through to the destination.
     delete req.headers['transfer-encoding'];
 
-    var dispatcher = function(req, res) {
-      req.method == 'CONNECT' ? handleSecure(req, res) : handlePlain(req, res);
-    }
-
     if (options.verbose) logRequest(req);
+    if (typeof self._logHandler == 'object') self._logHandler.log(socket, req);
 
-    if(typeof self._logHandler == 'object') {
-      self._logHandler.log(socket, req);
+    if (typeof self._authHandler != 'object') return acceptRequest();
+
+    // perform basic proxy auth (over established SSL tunnel)
+    // - http://www.chromium.org/spdy/spdy-authentication
+    var header = req.headers['proxy-authorization'];
+    if (!header) return rejectRequest('Proxy Authentication Required');
+    var tokens = header.match(/^(\S+)\s+(\S+)\s*$/);
+    if (!tokens) return rejectRequest('Proxy Authentication Header Malformed');
+    if (tokens[1].toLowerCase() != 'basic') return rejectRequest('Proxy Authentication Scheme Not Supported');
+    var userpass = new Buffer(tokens[2], 'base64').toString().match(/^([^:]*):(.*)$/);
+    if (!userpass) return rejectRequest('Malformed Credential for HTTP Basic Auth');
+    var username = userpass[1], password = userpass[2];
+
+    // don't pass proxy-auth headers upstream
+    delete req.headers['proxy-authorization'];
+
+    self._authHandler.authUser(username, password, function(authPassed) {
+      if (authPassed) {
+        return acceptRequest();
+      } else {
+        return rejectRequest();
+      }
+    });
+
+    function acceptRequest() {
+      if (https) {
+        handleSecure(req, res);
+      } else {
+        handlePlain(req, res);
+      }
     }
 
-    if(typeof self._authHandler == 'object') { // an AuthHandler is defined
-      // perform basic proxy auth (over established SSL tunnel)
-      // - http://www.chromium.org/spdy/spdy-authentication
-      var header = req.headers['proxy-authorization'] || '',
-          token = header.split(/\s+/).pop() || '',
-          auth = new Buffer(token, 'base64').toString(),
-          parts = auth.split(/:/),
-          username = parts[0],
-          password = parts[1];
+    function rejectRequest(err) {
+      err = err || 'Proxy Authentication Failed';
 
-      // don't pass proxy-auth headers upstream
-      delete req.headers['proxy-authorization'];
-
-      self._authHandler.authUser(username, password, function(authPassed) {
-        if (authPassed)
-          return dispatcher(req, res);
-
-        synReply(socket, 407, 'Proxy Authentication Required',
-          {'proxy-authenticate': 'Basic realm="SPDY Proxy"'},
-          function() {
-            socket.end();
-          }
-        );
-      });
-    } else { // auth is not necessary, simply go ahead and dispatch to funcs
-      dispatcher(req, res);
+      synReply(socket, 407, err,
+        {'Proxy-Authenticate': 'Basic realm="SPDY Proxy"'},
+        function() {
+          socket.end();
+        }
+      );
     }
-
   }
 
   spdy.server.Server.call(this, options);
